server/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
python-multipart==0.0.9
aiofiles==24.1.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
email-validator>=2.1.0
gunicorn==21.2.0
slowapi==0.1.9
python-dotenv==1.0.1
requests==2.32.3
